28/07/25

+++ src/main/java/com/erdal/model/SaloonReport.java
@@ -11,7 +11,7 @@ public class SaloonReport {
 
 	private Double totalEarnings;
 
-	private Integer totalBookings;
+	private Double totalBookings;
 
 	private Double totalRefund;
